Split parse_packet into separate file handle_packet.c
